Add Page controller
